requirements.txt
loguru
freesms
ruamel.yaml
requests
websockets
urllib3
aiohttp